requirements.txt
openai
python-dotenv
rich
pandas
chromadb
sentence-transformers
flask
